+++ src/brieflz_hashchain.h
@@ -3,7 +3,7 @@
 //
 // Lazy parsing with chains of previous positions
 //
-// Copyright (c) 2016-2018 Joergen Ibsen
+// Copyright (c) 2016-2020 Joergen Ibsen
 //
 // This software is provided 'as-is', without any express or implied
 // warranty. In no event will the authors be held liable for any damages
@@ -28,8 +28,8 @@
 #ifndef BRIEFLZ_HASHCHAIN_H_INCLUDED
 #define BRIEFLZ_HASHCHAIN_H_INCLUDED
 
-static unsigned long
-blz_hashchain_workmem_size(unsigned long src_size)
+static size_t
+blz_hashchain_workmem_size(size_t src_size)
 {
 	return (LOOKUP_SIZE + src_size) * sizeof(unsigned long);
 }
