+++ src/brieflz_sliding.h
@@ -3,7 +3,7 @@
 //
 // Lazy parsing with a sliding window of previous positions
 //
-// Copyright (c) 2016-2018 Joergen Ibsen
+// Copyright (c) 2016-2020 Joergen Ibsen
 //
 // This software is provided 'as-is', without any express or implied
 // warranty. In no event will the authors be held liable for any damages
@@ -28,8 +28,8 @@
 #ifndef BRIEFLZ_SLIDING_H_INCLUDED
 #define BRIEFLZ_SLIDING_H_INCLUDED
 
-static unsigned long
-blz_sliding_workmem_size(unsigned long src_size, int window_log)
+static size_t
+blz_sliding_workmem_size(size_t src_size, int window_log)
 {
 	(void) src_size;
